🚚 JE-36-chore: review 폴더 생성 및 구분
