+++ src/components/ActivityDetail/index.tsx
@@ -3,4 +3,4 @@ export { default as BannerImage } from './BannerImage';
 export { default as MobileBannerImage } from './MobileBannerImage';
 export { default as Description } from './Description';
 export { default as Address } from './Address';
-export { default as Review } from './Review';
+export { default as Review } from './Reivew/Review';
